fixed major bugs

+++ api/logbook.php
@@ -6,35 +6,41 @@ function createActivity(){
 		$conn=connect_db();
 
 		if(
-			isset($_POST['day_no']) &&  !empty($_POST['day_no']) &&
+			
 			isset($_POST['activity_details']) && !empty($_POST['activity_details']) &&
 			isset($_POST['student_id']) && !empty($_POST['student_id'])
 		){
-			$day_no=mysqli_real_escape_string($conn,$_POST['day_no']);
+			
 			$activity_details=mysqli_real_escape_string($conn,$_POST['activity_details']);
 			$student_id=mysqli_real_escape_string($conn,$_POST['student_id']);
 			
 
-			$sql = "INSERT INTO logbook ";
-			$sql .= "(day_no, activity_details, student_id, date_created, approved ";
-			$sql .= ") VALUES ('$day_no', '$activity_details', '$student_id', ";
-			$sql .= "CURRENT_DATE(), '0')";
 
-			$result = mysqli_query($conn, $sql);
+					
 
-			if (!$result) {
-				echo json_encode(array(
-					'status' => 'error',
-					'message' => mysqli_error($conn)
-				));
-				exit();
-			} else {
-				echo json_encode(array(
-					'status' => 'success',
-					'message' => 'Activity added'
-				));
-				exit();
-			}
+					$sql = "INSERT INTO logbook ";
+					$sql .= "(activity_details, student_id, date_created, approved ";
+					$sql .= ") VALUES ('$activity_details', '$student_id', ";
+					$sql .= "CURRENT_DATE(), '0')";
+
+					$result = mysqli_query($conn, $sql);
+
+					if (!$result) {
+						echo json_encode(array(
+							'status' => 'error',
+							'message' => mysqli_error($conn)
+						));
+						exit();
+					} else {
+						echo json_encode(array(
+							'status' => 'success',
+							'message' => 'Activity added'
+						));
+						exit();
+					}
+
+			
+			
 		}
 		else{
 			echo json_encode(array(
@@ -45,6 +51,9 @@ function createActivity(){
 		}
 
 	}
+
+
+
 // activity by id
     	function activityById($id=''){
 		$conn=connect_db();
@@ -79,9 +88,17 @@ function createActivity(){
 	}
 
 
+
+
 //delete activity
-function deleteActivity($id=''){
+function deleteActivity($id='', $dateCreated=""){
+
 		$conn=connect_db();
+
+	$toDate=date("Y-m-d");
+
+	if($dateCreated==$toDate){
+		//continue to delete
 		$sql = "DELETE FROM logbook WHERE logbook.id = $id";
 		$result = mysqli_query($conn, $sql);
 		if (!$result) {
@@ -101,6 +118,21 @@ function deleteActivity($id=''){
 				exit();
 			
 		}
+	}
+	else{
+		echo json_encode(array(
+				'status' => 'error',
+				'message' => 'Cannot Delete an activity from the past '.$toDate."---".$dateCreated
+			));
+			exit();
+	}
+
+
+
+
+
+
+		
 	}
 
 function approveActivity($acId=''){
@@ -132,7 +164,7 @@ function approveActivity($acId=''){
 // get progress
 	   	function loogBookProgress($id=''){
 		$conn=connect_db();
-		$sql = "SELECT day_no FROM logbook WHERE student_id=$id GROUP BY day_no";
+		$sql = "SELECT date_created FROM logbook WHERE student_id=$id GROUP BY 	date_created";
 		$result = mysqli_query($conn, $sql);
 		if (!$result) {
 			
@@ -166,9 +198,10 @@ function approveActivity($acId=''){
 	}
 
 //get activity by day
-    	function activityByDay($id='', $st=""){
+    	function activityByDay($dc='', $st=""){
+			
 		$conn=connect_db();
-		$sql = "SELECT * FROM logbook WHERE day_no = $id AND student_id=$st";
+		$sql = "SELECT * FROM logbook WHERE date_created ='$dc' AND student_id=$st";
 		$result = mysqli_query($conn, $sql);
 		if (!$result) {
 			
@@ -191,7 +224,7 @@ function approveActivity($acId=''){
 				
 				echo json_encode(array(
 					'status' => 'failed',
-					'message' => 'There are no activities on the day'
+					'message' => 'There are no activities on the day '.$sql
 				));
 				exit();
 			}
